refactor: group components and update imports
