+++ src/app/(auth)/login/page.jsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
-import Button from "../../components/Button";
+import Button from "../../components/buttons/Button";
 import Loading from "../../loading";
 import "./LoginPage.css";
 
@@ -42,12 +42,12 @@ export default function LoginPage() {
       if (!res.ok) {
         throw new Error("Invalid username or password. Please try again.");
       }
-
       const data = await res.json();
 
       // If login is successful, store authentication status and navigate to home
       localStorage.setItem("isAuth", JSON.stringify(true));
       localStorage.setItem("accessToken", data.accessToken);
+
       router.push("/home");
     } catch (error) {
       console.error(error);
