+++ src/app/(dashboard)/blog/BlogItem.js
@@ -1,9 +1,16 @@
 // Blog item component
-import Button from "../../components/Button";
+import Button from "../../components/buttons/Button";
 import Link from "next/link";
-import DeleteButton from "../../components/DeleteButton";
-import EditButton from "../../components/EditButton"
-export default function BlogItem({ id, title, content, views, onDelete, onEdit }) {
+import DeleteButton from "../../components/buttons/DeleteButton";
+import EditButton from "../../components/buttons/EditButton";
+export default function BlogItem({
+  id,
+  title,
+  content,
+  views,
+  onDelete,
+  onEdit,
+}) {
   return (
     <li className="blog__list__item">
       <div className="blog__content">
@@ -26,7 +33,7 @@ export default function BlogItem({ id, title, content, views, onDelete, onEdit }
           <Link className="blog__link" href={`/blog/${id}`}>
             <Button className="btn blog-btn" name="Read more" />
           </Link>
-          <EditButton onEdit={onEdit} id={id}/>
+          <EditButton onEdit={onEdit} id={id} />
           <DeleteButton onDelete={onDelete} />
         </div>
       </div>
