+++ src/app/(dashboard)/blog/[id]/PostDetails.css
@@ -1,3 +1,15 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 10rem;
+  width: 100%;
+  max-width: 144rem;
+  margin: 0 auto;
+  padding: 0 4rem;
+}
+
 .post-details-wrapper {
   display: flex;
   flex-direction: column;
