+++ src/app/(dashboard)/contact/page.js
@@ -1,5 +1,5 @@
 import "./Contact.css";
-import Button from "../../components/Button";
+import Button from "../../components/buttons/Button";
 
 function Contact() {
   return (
