+++ src/app/components/functions/handleDelete.js
